COMP: Remove hidden virtual function warnings
/Volumes/scratch/src/BSA-Darwin-clang/OpenCV-install/include/opencv2/ml/ml.hpp:957:18: warning: 'CvForestTree::train' hides overloaded virtual functions [-Woverloaded-virtual]
    virtual bool train( CvDTreeTrainData* trainData, const CvMat* _subsample_idx, CvRTrees* forest );
                 ^
/Volumes/scratch/src/BSA-Darwin-clang/OpenCV-install/include/opencv2/ml/ml.hpp:861:18: note: hidden overloaded virtual function 'CvDTree::train' declared here
    virtual bool train( CvMLData* trainData, CvDTreeParams params=CvDTreeParams() );
                 ^

The derived classes for CvForestTree and CvBoostTree have specific train function signatures
that do not exist in the parent class.  All the parent class train function
signatures need to be made invalid because they should never be used.

In order to accomplish this without compiler warnings, each of the parent class train function
signatures must be explicitly declared in the child class, and defined to simply abort();